+++ src/main/java/de/maxhenkel/voicechat/voice/common/Utils.java
@@ -1,8 +1,7 @@
 package de.maxhenkel.voicechat.voice.common;
 
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.settings.PointOfView;
-import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.client.renderer.ActiveRenderInfo;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.vector.Vector2f;
 import net.minecraft.util.math.vector.Vector3d;
@@ -118,13 +117,13 @@ public class Utils {
 
     @OnlyIn(Dist.CLIENT)
     public static Pair<Float, Float> getStereoVolume(Minecraft minecraft, Vector3d soundPos, double voiceChatDistance) {
-        PlayerEntity player = minecraft.player;
-        Vector3d playerPos = player.position();
-        Vector3d d = soundPos.subtract(playerPos).normalize();
+        ActiveRenderInfo mainCamera = minecraft.gameRenderer.getMainCamera();
+        Vector3d cameraPos = mainCamera.getPosition();
+        Vector3d d = soundPos.subtract(cameraPos).normalize();
         Vector2f diff = new Vector2f((float) d.x, (float) d.z);
         float diffAngle = angle(diff, new Vector2f(-1F, 0F));
-        float angle = normalizeAngle(diffAngle - (player.yRot % 360F));
-        float dif = (float) (Math.abs(playerPos.y - soundPos.y) / voiceChatDistance);
+        float angle = normalizeAngle(diffAngle - (mainCamera.getYRot() % 360F));
+        float dif = (float) (Math.abs(cameraPos.y - soundPos.y) / voiceChatDistance);
 
         float rot = angle / 180F;
         float perc = rot;
@@ -141,10 +140,6 @@ public class Utils {
         float fill = 1F - Math.max(left, right);
         left += fill;
         right += fill;
-
-        if (minecraft.options.getCameraType().equals(PointOfView.THIRD_PERSON_FRONT)) {
-            return new ImmutablePair<>(right, left);
-        }
 
         return new ImmutablePair<>(left, right);
     }